feeder/requirements.txt
attrs==15.2.0
cffi==1.5.2
cryptography==1.2.2
cssselect==0.9.1
dj-database-url==0.4.0
Django==1.8
django-allauth==0.25.2
django-cors-headers==1.1.0
django-rest-auth==0.7.0
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
enum34==1.1.2
feedparser==5.2.1
gunicorn==19.4.5
idna==2.0
ipaddress==1.0.16
lxml==3.5.0
oauthlib==1.0.3
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyJWT==1.4.0
pyOpenSSL==0.15.1
python-openid==2.2.5
queuelib==1.4.2
requests==2.9.1
requests-oauthlib==0.6.1
Scrapy==1.0.5
scrapy-djangoitem==1.0.0
service-identity==16.0.0
six==1.10.0
Twisted==15.5.0
w3lib==1.13.0
wheel==0.26.0
whitenoise==2.0.6
zope.interface==4.1.3
